% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianssd-package.R
\docType{package}
\name{bayesianssd}
\alias{bayesianssd-package}
\alias{bayesianssd}
\title{bayesianssd: Bayesian Sample Size Determination}
\description{
Provides tools for Bayesian sample size determination in clinical trials
and other experimental settings. Includes functions for simulation-based
design, posterior analysis, and decision criteria.
}
\details{
This package implements methods for Bayesian adaptive design and
sample size justification using prior and posterior predictive approaches.
}
\seealso{
\code{\link{runSSD}},
\link{https://github.com/GitCJW/bayesianssd}
}
\author{
Christoph Waterkamp \email{christoph.waterkamp@uni-due.de}
Jeannine Veseljak
}
