% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{checkSettings}
\alias{checkSettings}
\title{Input validity check}
\usage{
checkSettings(model, dataCreationFunction, minN, goals)
}
\arguments{
\item{model}{An object of class \code{stanmodel, stanreg or brmsfit}.}

\item{dataCreationFunction}{A function that accepts a single parameter,
\code{N}, and should generates N values in the same manner as the given \code{model}.}

\item{minN}{The minimum number of data points. If your data contains e.g. 4
groups the minimum number should be at least 4.}

\item{goals}{The condition be tested. Use the function
\link[bayesianssd]{createGoal}  to create such goals.}
}
\value{
Logical scalar. `TRUE` if all required inputs are valid and consistent.
}
\description{
Checks if the input for a call to \code{run_sdd} is valid.
}
\examples{
\dontshow{if (rlang::is_installed("rstanarm")) withAutoprint(\{ # examplesIf}
dataCreationFunction <- function(N){
 group_effects <- c(
   control = 13,
   drug = 7
 )
 treatment <- rep(c("control", "drug"), length.out=N)
 y <- rpois(N, group_effects)

 data <- data.frame(
   treatment = treatment,
   y = y
 )
 data
}

model <- rstanarm::stan_glm("y~-1+treatment", data=dataCreationFunction(20), family=poisson())

goal <- createGoal(parametersA = "treatmentcontrol", parametersB = "treatmentdrug",
                  goalType = "rope", ropeType = "exclude", ropeLower = 0, ropeUpper = 0, ci = 0.95)

checkSettings(model, dataCreationFunction, 2, list(goal))
\dontshow{\}) # examplesIf}
}
