% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goal.R
\name{createGoal}
\alias{createGoal}
\title{Single goal for a sample size determination}
\usage{
\code{createGoal(
  parametersA = c(),
  parametersB = NULL,
  goalType = c("rope","precision"),
  ci = 0.95,
  ropeType = c("exclude","include"),
  ropeLower = NULL,
  ropeUpper = NULL,
  ropeExclusive = T,
  precisionWidth = NULL
)}
}
\arguments{
\item{parametersA}{A vector of parameters defining the first 'group' of posteriors.}

\item{parametersB}{A vector of parameters defining the second 'group' of posteriors.
The difference of the two posterior sums defined by \code{parametersA} and \code{parametersB} (if set, otherwise 0) is the target posterior.}

\item{goalType}{One of the two goal types \code{"rope", "precision"}.}

\item{ci}{The target credible interval of the posterior.}

\item{ropeType}{One of the two rope types \code{"exclude", "include"}.
For \code{ropeType="exclude"} the credible interval must be outside the ROPE.
If \code{ropeType="include"}, the credible interval must be inside the ROPE.
Used only when \code{goalType="rope"}.}

\item{ropeLower}{Lower bound of the ROPE. Used only when \code{goalType="rope"}.}

\item{ropeUpper}{Upper bound of the ROPE. Used only when \code{goalType="rope"}.}

\item{ropeExclusive}{Used only for a ROPE and \code{'exclude'}, default TRUE.
If set to TRUE: When the posterior extends beyond both bounds of the ROPE, the larger fraction of the posterior outside the ROPE is used.
If set to FALSE: The difference between the two fractions outside the ROPE is calculated.}

\item{precisionWidth}{The maximum allowed width of the credible interval. Used only when \code{goalType="precision"}.}
}
\value{
A goal that can be passed to \code{runSSD}.
}
\description{
Creates a single goal for a sample size determination passed to \code{runSSD}.
}
\examples{
\code{
createGoal(
parametersA="mu[1]", parametersB="mu[2]",
goalType="rope", ropeType="exclude",
ropeLower=-0.1, ropeUpper = 0.1, ropeExclusive=T,
ci=0.95)}
}
