% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print.R
\name{format.bayesianssd}
\alias{format.bayesianssd}
\title{Summary of sample size determination result}
\usage{
\method{format}{bayesianssd}(ssd)
}
\arguments{
\item{ssd}{An object of class `"bayesianssd"` returned by \code{runSSD()}.}
}
\value{
A character scalar containing the textual summary of the sample size
determination.
}
\description{
Summary of sample size determination result
}
\examples{
\dontshow{if (rlang::is_installed("rstanarm")) withAutoprint(\{ # examplesIf}
\dontest{
dataCreationFunction <- function(N){
 group_effects <- c(
   control = 13,
   drug = 7
 )
 treatment <- rep(c("control", "drug"), length.out=N)
 y <- rpois(N, group_effects)

 data <- data.frame(
   treatment = treatment,
   y = y
 )
 data
}

model <- stan_glm("y~-1+treatment", data=dataCreationFunction(20), family=poisson())

goal <- createGoal(parametersA = "treatmentcontrol", parametersB = "treatmentdrug",
                  goalType = "rope", ropeType = "exclude", ropeLower = 0, ropeUpper = 0, ci = 0.95)

checkSettings(model, dataCreationFunction, 2, list(goal))

plan("multisession")
ssd <- runSSD(
 model = model,
 dataCreationFunction = dataCreationFunction,
 powerDesired = 0.8,
 minN = 2,
 maxN = 20,
 goals = list(goal),
 con = 200,
 iParallel = 20)

summary <- format(ssd)
print(summary)
}
\dontshow{\}) # examplesIf}
}
