% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goal.R
\name{goalList}
\alias{goalList}
\title{Goal list for a sample size determination}
\usage{
goalList(...)
}
\arguments{
\item{...}{Object(s) of  class `"bayesianssdgoal"`.}
}
\value{
An Object of class `"bayesianssdgoallist"` that can be passed to \code{plot}.
}
\description{
Creates a list of goals for a sample size determination passed to \code{runSSD}.
}
\examples{
goal1 <- createGoal(
parametersA="mu[1]", parametersB="mu[2]",
goalType="rope", ropeType="exclude",
ropeLower=-0.1, ropeUpper = 0.1, ropeExclusive=TRUE,
ci=0.95)

goal2 <- createGoal(
parametersA="mu[1]", parametersB="mu[3]",
goalType="rope", ropeType="exclude",
ropeLower=-0.1, ropeUpper = 0.1, ropeExclusive=TRUE,
ci=0.95)

listOfGoals <- goalList(goal1, goal2)
}
