% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot.bayesianssd}
\alias{plot.bayesianssd}
\title{Plots the sample size determination}
\usage{
\method{plot}{bayesianssd}(
  ssd,
  plotTriangles = TRUE,
  theme = list(main = "#56B4E9", current = "black", default = "darkgrey", mean = "white")
)
}
\arguments{
\item{ssd}{Object returned by \code{runSSD()}.}

\item{plotTriangles}{Boolean (default: TRUE) indicating whether to draw
triangles representing the maximum accepted power width}

\item{theme}{A list given the colors for the plot.
Default: \code{list(main="#56B4E9", current="black", default="darkgrey",
mean="white")}.}
}
\value{
A ggplot of the final sample size determination.
}
\description{
Plots the sample size determination
}
\examples{
\dontshow{if (rlang::is_installed("rstanarm") && interactive()) withAutoprint(\{ # examplesIf}
\dontest{
dataCreationFunction <- function(N){
 group_effects <- c(
   control = 13,
   drug = 7
 )
 treatment <- rep(c("control", "drug"), length.out=N)
 y <- rpois(N, group_effects)

 data <- data.frame(
   treatment = treatment,
   y = y
 )
 data
}

model <- stan_glm("y~-1+treatment", data=dataCreationFunction(20), family=poisson())

goal <- createGoal(parametersA = "treatmentcontrol", parametersB = "treatmentdrug",
                  goalType = "rope", ropeType = "exclude", ropeLower = 0, ropeUpper = 0, ci = 0.95)

checkSettings(model, dataCreationFunction, 2, list(goal))

plan("multisession")
ssd <- runSSD(
 model = model,
 dataCreationFunction = dataCreationFunction,
 powerDesired = 0.8,
 minN = 2,
 maxN = 20,
 goals = list(goal),
 con = 200,
 iParallel = 20)

plot(ssd)
}
\dontshow{\}) # examplesIf}
}
