% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot.bayesianssdgoal}
\alias{plot.bayesianssdgoal}
\title{Plot goals for a sample size determination}
\usage{
\method{plot}{bayesianssdgoal}(x, dataCreationFunction = NULL, model = NULL, N = NULL)
}
\arguments{
\item{x}{The condition be tested. Use the function
\link[bayesianssd]{createGoal} to create such goals.}

\item{dataCreationFunction}{A function that accepts a single parameter,
\code{N}, and generates N values in the same manner as the given \code{model}.}

\item{model}{A object of class \code{stanmodel, stanreg or brmsfit}.}

\item{N}{The exemplary sample size.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots a single goal created by \code{createGoal}.
If the additional arguments are provided, the goals are plotted alongside example data.
}
\examples{
goal <- createGoal(parametersA="treatmentcontrol", parametersB="treatmentdrug",
                   goalType="rope", ropeType="include", ropeLower=-2, ropeUpper=2,
                   ropeExclusive=TRUE, ci=0.95)
plot(goal)


\dontshow{if (rlang::is_installed("rstanarm")) withAutoprint(\{ # examplesIf}
dataCreationFunction <- function(N){
  group_effects <- c(
    control = 9,
    drug = 7
  )
  treatment <- rep(c("control", "drug"), length.out=N)
  y <- rpois(N, group_effects)

  data <- data.frame(
    treatment = treatment,
    y = y
  )
  data
}

model <- rstanarm::stan_glm("y~-1+treatment", data=dataCreationFunction(20), family=poisson())

goal <- createGoal(parametersA="treatmentcontrol", parametersB="treatmentdrug",
                   goalType="rope", ropeType="exclude", ropeLower=-0.1, ropeUpper = 0.1,
                   ropeExclusive=TRUE, ci=0.95)
plot(goal, dataCreationFunction, model, N=100)
\dontshow{\}) # examplesIf}
}
