% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plotResults}
\alias{plotResults}
\title{Plots the sample size determination}
\usage{
plotResults(
  ssd,
  plotTriangles = T,
  theme = list(main = "#56B4E9", current = "black", default = "darkgrey", mean = "white")
)
}
\arguments{
\item{ssd}{Object returned by \code{runSSD()}.}

\item{plotTriangles}{Boolean (default: TRUE) indicating whether to draw triangles representing the maximum accepted power width}

\item{theme}{A list given the colors for the plot. Default: \code{list(main="#56B4E9", current="black", default="darkgrey", mean="white")}.}
}
\value{
A ggplot.
}
\description{
Plots the sample size determination
}
