% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{printSSD}
\alias{printSSD}
\title{Prints sample size determination result}
\usage{
printSSD(ssd, summarized = T)
}
\arguments{
\item{ssd}{Object returned by \code{runSSD()}.}

\item{summarized}{Boolean (default: TRUE) whether a summarized version of all steps should be printed.}
}
\description{
Prints sample size determination result
}
