% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{checkSettings}
\alias{checkSettings}
\title{Input validity check}
\usage{
checkSettings(model, dataCreationFunction, minN, goals)
}
\arguments{
\item{model}{An object of class \code{stanmodel, stanreg or brmsfit}.}

\item{dataCreationFunction}{A function that accepts a single parameter, \code{N}, and should generates N values in the same manner as the given \code{model}.}

\item{minN}{The minimum number of data points. If your data contains e.g. 4 groups the minimum number should be at least 4.}

\item{goals}{The condition be tested. Use the function \link[bayesianssd]{createGoal}  to create such goals.}
}
\value{
TRUE if everything seems to be well defined.
}
\description{
Checks if the input for a call to \code{run_sdd} is valid.
}
